rename(all): domain 아래 ladder, player 패키지 추가
