+++ src/main/java/ladder/domain/player/PlayerGroup.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.player;
 
 import java.util.*;
 import java.util.stream.Collectors;
