+++ src/test/java/ladder/domain/ladder/LadderRowTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.ladder;
 
-import ladder.domain.LadderRow;
 import ladder.strategy.LineStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
