+++ src/test/java/ladder/domain/ladder/LadderTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.ladder;
 
-import ladder.domain.Ladder;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
