+++ src/test/java/ladder/domain/player/PlayerGroupTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.player;
 
-import ladder.domain.PlayerGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
