+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,6 +1,5 @@
 package ladder.domain.player;
 
-import ladder.domain.Player;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
